requirements.txt
Pillow>=10.2.0
pynput>=1.7.6
psutil>=5.9.8
pywinauto>=0.6.8
PyYAML>=6.0.1
requests>=2.32.3
mss>=9.0.1
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
numpy>=2.2.4
pymysql>=1.1.1
